George v.4
Got everything to work, atleast without user restriction to how many session they have. I added a sessions tab and made it possible to create different sessions with the users. Haven't tested it with users other than admin, so I will do that next. If this works what we need to do:
-make user restrictions
-other games
-other versions
-Better User Interface